Crate gourd

gourd

gourd is a command-line tool that schedules parallel runs for algorithm comparisons.

Given the parameters of the experiment, a number of test datasets, and algorithm implementations to compare, gourd runs the experiment in parallel and provides many options for processing its results.

While originally envisioned for the DelftBlue supercomputer at Delft University of Technology, gourd can replicate the experiment on any cluster computer with the Slurm scheduler, on any UNIX-like system, and on Microsoft Windows.

Usage

Verify that gourd is installed by running:

$ gourd version

Congratulations! Now you can use gourd.

There is extensive tutorials and documentation which can be accessed in many different formats.

As a manpage, with the man command:

$ man gourd
$ man gourd.toml
$ man gourd-tutorial

Modules

analyse

Analysing runs - collecting metrics, exporting, plotting.

chunks

An implementation for allocating queued jobs to chunks

cli

The command line interface and relevant structures.

experiments

Extensions to the Experiment struct defined in gourd-lib, allowing for operations on runtime data.

init

Functionality for user-friendly initialisation of new experimental setups and examples.

local

A framework for running experiments on the local machine using a thread-pool executor.

post

Functionality for post-processing jobs including after-scripts, pipeline jobs, and retrieval of their status.

rerun

Rerun subcommand helper functions

slurm

A framework for running supercomputer experiments by interfacing with a local installation of SLURM.

status

Functionality for checking and displaying the status of already running experiments.

wrapper

An interface for communicating with gourd-wrapper, a separately packaged binary that encapsulates the user’s programs.

Functions

main 🔒

The main CLI entry-point of the gourd utility.