gourd

Module analyse

Analysing runs - collecting metrics, exporting, plotting.

Modules

csvs

Export experiment data as CSV file

plotting

Draw up plots of experiment data

Structs

Column

A column that can be appended to the end of a Table.

ColumnGenerator

Create a Column from a list of entries of type X.

Table

Represent a human-readable table. Universal between CSV exporting and in-line display.

Functions

get_completion_time

Get completion time of a run.

get_completions

Get completion times of jobs.