gourd::chunks

Trait Chunkable

pub trait Chunkable {
    // Required methods
    fn next_chunks(
        &mut self,
        chunk_length: usize,
        how_many: usize,
        status: ExperimentStatus,
    ) -> Result<Vec<Chunk>>;
    fn register_runs(&mut self, runs: &[usize]) -> usize;
    fn mark_chunk_scheduled(&mut self, chunk: &Chunk, batch_id: String);
    fn unscheduled(&self, status: &ExperimentStatus) -> Vec<(usize, &Run)>;
    fn scheduled_nodep(&self) -> usize;
}

Split an Experiment’s Runs into Chunks of common ResourceLimits.

Required Methods

fn next_chunks( &mut self, chunk_length: usize, how_many: usize, status: ExperimentStatus, ) -> Result<Vec<Chunk>>

Next available Chunks for scheduling,

fn register_runs(&mut self, runs: &[usize]) -> usize

Add the runs to the experiment so the wrapper can find them,

returns the chunk index that was just created.

fn mark_chunk_scheduled(&mut self, chunk: &Chunk, batch_id: String)

Once a chunk has been scheduled, mark all of its runs as scheduled with their slurm ids

fn unscheduled(&self, status: &ExperimentStatus) -> Vec<(usize, &Run)>

Get the still pending runs of this experiment.

fn scheduled_nodep(&self) -> usize

Get the still pending runs of this experiment.