gourd::experiments

Trait ExperimentExt

pub trait ExperimentExt {
    // Required methods
    fn from_config(
        conf: &Config,
        time: DateTime<Local>,
        env: Environment,
        fs: &impl FileOperations,
    ) -> Result<Self>
       where Self: Sized;
    fn latest_id_from_folder(folder: &Path) -> Result<Option<usize>>;
    fn latest_experiment_from_folder(
        folder: &Path,
        fs: &impl FileOperations,
    ) -> Result<Experiment>;
    fn experiment_from_folder(
        seq: usize,
        folder: &Path,
        fs: &impl FileOperations,
    ) -> Result<Experiment>;
}

Extension trait for the shared Experiment struct.

Required Methods

fn from_config( conf: &Config, time: DateTime<Local>, env: Environment, fs: &impl FileOperations, ) -> Result<Self>
where Self: Sized,

Initialize a new experiment from a config.

Creates a new experiment by matching all algorithms to all inputs. The experiment is created in the provided env and with time as the timestamp.

fn latest_id_from_folder(folder: &Path) -> Result<Option<usize>>

Get the filename of the newest experiment.

fn latest_experiment_from_folder( folder: &Path, fs: &impl FileOperations, ) -> Result<Experiment>

Provided a folder gets the most recent experiment.

fn experiment_from_folder( seq: usize, folder: &Path, fs: &impl FileOperations, ) -> Result<Experiment>

Provided a folder gets the specified experiment.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.