gourd

Module init

Functionality for user-friendly initialisation of new experimental setups and examples.

Modules

builtin_examples

Tarballs of built-in example templates (optional feature)

interactive

Interactive configuration of an experiment setup template.

Functions

check_init_directory 🔒

Verifies that a directory is valid for the gourd init command.

init_experiment_setup

Initializes an experimental setup.

init_from_example 🔒

Retrieves and unpacks the example denoted by the provided example-id.

list_init_examples

Lists valid example IDs for this build of gourd init.