gourd::rerun

Module slurm

User interactions for SLURM reruns. This is interactive and SLURM-bound so cannot be checked locally.

Functions

check_multiple_runs_failed 🔒

Check the statuses of a list of runs and ask the user what to rerun.

check_single_run_failed 🔒

Check the status of a single run and ask the user what to rerun.

query_changing_limits_for_programs

Query the user for the resource limits of the programs for a list of runs.

query_changing_resource_limits

Ask the user if they want to change any resource limits for the current experiment.