gourd::slurm

Trait SlurmInteractor

pub trait SlurmInteractor {
    // Required methods
    fn get_version(&self) -> Result<[u64; 2]>;
    fn get_partitions(&self) -> Result<Vec<Vec<String>>>;
    fn max_array_size(&self) -> Result<usize>;
    fn max_submit(&self) -> Result<usize>;
    fn max_jobs(&self) -> Result<usize>;
    fn schedule_chunk(
        &self,
        slurm_config: &SlurmConfig,
        chunk: &Chunk,
        experiment: &mut Experiment,
        exp_path: &Path,
    ) -> Result<()>;
    fn is_version_supported(&self, v: [u64; 2]) -> bool;
    fn get_supported_versions(&self) -> String;
    fn get_accounting_data(
        &self,
        since: &DateTime<Local>,
    ) -> Result<Vec<SacctOutput>>;
    fn scheduled_jobs(&self) -> Result<Vec<String>>;
    fn scheduled_count(&self) -> Result<usize>;
    fn cancel_jobs(&self, batch_ids: Vec<String>) -> Result<()>;
}

The interface for interacting with a SLURM cluster. This can be via a version-specific CLI, via a REST API, or via a library.

Required Methods

fn get_version(&self) -> Result<[u64; 2]>

Check the version of slurm on the current environment. returns an error if the version is not supported, or if slurm is not present.

fn get_partitions(&self) -> Result<Vec<Vec<String>>>

Check if the provided partition is valid.

fn max_array_size(&self) -> Result<usize>

Get the MaxArraySize of this slurm cluster

fn max_submit(&self) -> Result<usize>

Get the MaxSubmit of this slurm cluster

fn max_jobs(&self) -> Result<usize>

Get the MaxJobs of this slurm cluster

fn schedule_chunk( &self, slurm_config: &SlurmConfig, chunk: &Chunk, experiment: &mut Experiment, exp_path: &Path, ) -> Result<()>

Schedule a new job array on the cluster.

fn is_version_supported(&self, v: [u64; 2]) -> bool

Check if a version of SLURM is supported by this interactor.

fn get_supported_versions(&self) -> String

Get the supported versions of SLURM for this interactor.

fn get_accounting_data( &self, since: &DateTime<Local>, ) -> Result<Vec<SacctOutput>>

Get accounting data of user’s jobs

fn scheduled_jobs(&self) -> Result<Vec<String>>

Get vector of all (not finished) jobs scheduled by user

fn scheduled_count(&self) -> Result<usize>

Get the number of currently scheduled or running jobs by the current user

fn cancel_jobs(&self, batch_ids: Vec<String>) -> Result<()>

Cancel all of the jobs in the batch_ids vector