gourd

Module status

Functionality for checking and displaying the status of already running experiments.

Modules

chunks

Printing information about scheduled chunks.

fs_based

File system based status information.

printing

Printing status information.

slurm_based

Slurm based status information.

slurm_files

Slurm file based status information.

Structs

FileSystemBasedStatus

Structure of file based status

SlurmBasedStatus

Structure of slurm based status

Status

The status of a single run. Contains FileSystemBasedStatus, and in case of running on slurm SlurmBasedStatus and SlurmFileOutput

Enums

FsState

This possible status of a job, reported by the file system.

SlurmState

The reasons for slurm to kill a job

Traits

DynamicStatus

Instead of storing a possibly outdated status somewhere, every time it is needed it’s fetched by the status module directly.

StatusProvider

A struct that can attest the statuses or some or all running jobs.

Functions

blocking_status

Print status until all tasks are finished.

merge_statuses

A function that merges status providers outputs.

Type Aliases

ExperimentStatus

This type maps between run_id and the Status of the run.