gourd::status

Module printing

Printing status information.

Functions

display_job

Display the status of an experiment in a human-readable from.

display_runs 🔒

Display runs of some group.

display_statuses

Display the status of an experiment in a human readable from.

format_input_name 🔒

For an input, decide how it’s shown to a user.

long_status 🔒

Display a shortened status for a small amount of runs.

short_status 🔒

Display a shortened status for a lot of runs.