gourd::status

Module slurm_based

Slurm based status information.

Structs

SacctOutput

Structure of Sacct output.

SlurmBasedProvider

Provide job status information based on the files system information.

Functions

flatten_job_id 🔒

This function takes SacctOutput and expands job ids.

flatten_slurm_id

Takes a job id as slurm spews out, eg 1234_[56-58] and expands it into [1234_56, 1234_57, 1234_58].